pyproject.toml
[project]
name = "rag-my-base"
version = "0.1.0"
description = "AI bot with some knowlage"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.23",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.3.2",
    "notebook>=7.4.2",
    "requests>=2.32.3",
]

[dependency-groups]
parser = [
    "beautifulsoup4>=4.13.4",
    "feedparser>=6.0.11",
    "langchain-chroma>=0.2.3",
    "langchain-text-splitters>=0.3.8",
]
